rename: 멤버 관련 상수값 member.ts로 이동
